backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
opencv-python
pytesseract
pdf2image
numpy
